+++ pkg/aws/awsmachinedeployment/validate_awsmachinedeployment_test.go
@@ -7,81 +7,10 @@ import (
 
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
-	"github.com/giantswarm/micrologger/microloggertest"
 
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/label"
-	"github.com/giantswarm/aws-admission-controller/v2/pkg/mutator"
 	"github.com/giantswarm/aws-admission-controller/v2/pkg/unittest"
 )
-
-func TestAWSMachineDeploymentAvailabilityZones(t *testing.T) {
-	testCases := []struct {
-		ctx  context.Context
-		name string
-
-		currentAZ     []string
-		expectedPatch []string
-	}{
-		{
-			// Don't default the AZ if they are set
-			name: "case 0",
-			ctx:  context.Background(),
-
-			currentAZ:     unittest.DefaultAvailabilityZones(),
-			expectedPatch: nil,
-		},
-		{
-			// Default the AZ they are not set
-			name: "case 1",
-			ctx:  context.Background(),
-
-			currentAZ:     nil,
-			expectedPatch: []string{unittest.DefaultMasterAvailabilityZone},
-		},
-	}
-	for i, tc := range testCases {
-		t.Run(strconv.Itoa(i), func(t *testing.T) {
-			var err error
-			var updatedAZs []string
-
-			fakeK8sClient := unittest.FakeK8sClient()
-			mutate := &Mutator{
-				k8sClient: fakeK8sClient,
-				logger:    microloggertest.New(),
-			}
-			awsControlPlane := unittest.DefaultAWSControlPlane()
-			err = fakeK8sClient.CtrlClient().Create(tc.ctx, &awsControlPlane)
-			if err != nil {
-				t.Fatal(err)
-			}
-
-			// run mutate function to default AWSMachineDeployment AZs
-			var patch []mutator.PatchOperation
-			awsmachinedeployment := unittest.DefaultAWSMachineDeployment()
-			awsmachinedeployment.Spec.Provider.AvailabilityZones = tc.currentAZ
-			patch, err = mutate.MutateAvailabilityZones(awsmachinedeployment)
-			if err != nil {
-				t.Fatal(err)
-			}
-			// parse patches
-			for _, p := range patch {
-				if p.Path == "/spec/provider/availabilityZones" {
-					updatedAZs = p.Value.([]string)
-				}
-			}
-
-			// check if the AZs are patched as expected
-			if len(tc.expectedPatch) != len(updatedAZs) {
-				t.Fatalf("expected %v to not to differ from %v", len(tc.expectedPatch), len(updatedAZs))
-			}
-			for i, p := range updatedAZs {
-				if tc.expectedPatch[i] != p {
-					t.Fatalf("expected %v to not to differ from %v", tc.expectedPatch[i], p)
-				}
-			}
-		})
-	}
-}
 
 func TestMachineDeploymentLabelMatch(t *testing.T) {
 	testCases := []struct {
